pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "job_search_tool"
version = "0.1.0"
authors = [
  { name="Jules", email="jules@example.com" },
]
description = "A tool to automate the discovery and pre-screening of freelance jobs."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "sqlalchemy",
    "click",
    "pyyaml",
    "beautifulsoup4",
    "requests",
]

[project.optional-dependencies]
dev = [
    "pytest",
]

[project.scripts]
jobtool = "job_search_tool.main:cli"

[tool.setuptools.packages.find]
where = ["src"]
